fix(projet): incorrect type

+++ projet/lib/include/Adc.h
@@ -29,7 +29,7 @@
 class Adc {
 public:
     static constexpr float VOLTAGE_REFERENCE = 5.0F;
-    static constexpr uint8_t MAX_ADC_VALUE = 1024;
+    static constexpr uint16_t MAX_ADC_VALUE = 1024;
 
     Adc(Adc&) = delete;
     void operator=(const Adc&) = delete;
